Eliminate utility class
